move UIMessage into types
