+++ src/react/index.ts
@@ -1,13 +1,11 @@
 "use client";
 import { useMemo, useState } from "react";
-import type { UIMessage } from "./toUIMessages.js";
-import { toUIMessages } from "./toUIMessages.js";
-import type { ThreadQuery } from "./types.js";
+export { toUIMessages } from "./toUIMessages.js";
+export type { ThreadQuery, UIMessage } from "./types.js";
 
 export { optimisticallySendMessage } from "./optimisticallySendMessage.js";
 export { useSmoothText } from "./useSmoothText.js";
 export { SmoothText } from "./SmoothText.js";
-export { toUIMessages, type UIMessage, type ThreadQuery };
 export {
   useThreadMessages,
   useStreamingThreadMessages,
